Moving models to models directory
