+++ packages/relative-values/.gitignore
@@ -14,6 +14,7 @@
 
 # production
 /build
+/models/cache/
 
 # misc
 .DS_Store
