+++ packages/relative-values/src/storage/StorageProvider.tsx
@@ -1,7 +1,7 @@
 import { generateProvider } from "@/components/generateProvider";
 import { InterfaceWithModels } from "@/types";
 import { FC, PropsWithChildren, Reducer } from "react";
-import { allInterfaces } from "@/builtins";
+import { allInterfaces } from "../../models/src";
 import { createEmptyGraphModel, Model } from "@/model/utils";
 
 type Storage = {
